+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/Mary/Translation.hs
@@ -13,7 +13,6 @@ import Cardano.Ledger.Binary
 import Cardano.Ledger.Core
 import Cardano.Ledger.Genesis (NoGenesis (..))
 import Cardano.Ledger.Mary (MaryEra)
-import Cardano.Ledger.Mary.Translation ()
 import qualified Cardano.Ledger.Shelley.API as S
 import Test.Cardano.Ledger.AllegraEraGen ()
 import Test.Cardano.Ledger.Binary.RoundTrip
@@ -54,7 +53,7 @@ maryTranslationTests =
   testGroup
     "Mary translation binary compatibiliby tests"
     [ testProperty "Tx compatibility" $
-        translateEraEncoding @MaryEra @S.ShelleyTx NoGenesis toCBOR toCBOR
+        translateEraEncoding @MaryEra @Tx NoGenesis toCBOR toCBOR
     , testProperty "ProposedPPUpdates compatibility" (test @S.ProposedPPUpdates)
     , testProperty "ShelleyGovState compatibility" $
         translateEraEncoding @MaryEra @S.ShelleyGovState NoGenesis toCBOR toCBOR